+++ jpostman-httpclient/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.0.3
+
+### Added
+- Added `RestAssuredExecutor.apply(RequestProvider)`.
+
 ## 1.0.2
 
 ### Added